[Source]: require source file to be utf8 encoded
Currently just scan the whole file upfront to ensure that it is utf8
encoded. The scan is performed in the Preprocessor.

Subsequent lexer enhancements will be needed for consuming unicode
codepoints instead of bytes (e.g. for identifiers); at that time the extra
scan can be removed and lexer.next() can either return an error or a special
token to indicate invalid utf8.