[Lexer]: extended identifier fixes
Subtracting codepoint_len from index is incorrect, since that might take
us back past the beginning of the buffer.

Instead, either update the state and set codepoint_len to zero
(so that the next iteration will try the same codepoint with a new
state); or set the token id directly and break, if there is only one
possible outcome.